+++ WalletWasabi.Fluent/ViewModels/StatusBar/StatusBarViewModel.cs
@@ -1,0 +1,7 @@
+namespace WalletWasabi.Fluent.ViewModels.StatusBar
+{
+	public class StatusBarViewModel
+	{
+		
+	}
+}
